requirements.txt
flask
flask-socketio
eventlet
# Hardware dependencies (will fail on Windows, handled in code)
spidev
RPi.GPIO
